+++ continuousintegration/docker/ctafrontend/cc7/opt/run/bin/kdc.sh
@@ -3,7 +3,9 @@
 . /opt/run/bin/init_pod.sh
 
 # Install missing RPMs (kdc)
-yum -y install heimdal-server heimdal-workstation
+if [ ! -e /etc/buildtreeRunner ]; then
+  yum -y install heimdal-server heimdal-workstation
+fi
 
 # Init the kdc store
 echo -n "Initing kdc... "
